Initial support for (TeX-style) Literate Haskell
Add a new option flag `--literate-tex` and pass it's presence down the call
stack to `transform`. When present, wrap the additional generated code in a
pair of `\begin{code}` and `\end{code}` delimiters.

I've opted for `--literate-tex` to not preclude the inclusion of a
`--literate-bird` for bird-style literate files in the future, although that
will require a lot more work because the pre-processing currently can't handle
those.